requirements.txt
glob2==0.6
fbprophet==0.4.post2
lightgbm==2.2.2
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
pylint==2.2.2
pytest==4.0.2
pyyaml==3.13
requests==2.21.0
scikit-learn==0.20.1
sklearn-pandas==1.8.0
Unidecode==1.0.23
xgboost==0.82
